requirements.txt
discord.py
dnspython
pyNaCl
async-timeout
tensorflow